PA-636
Removed and tested.

+++ projects/FieldNation/fieldnation/src/main/java/com/fieldnation/v2/ui/workorder/WorkOrderCard.java
@@ -621,7 +621,7 @@ public class WorkOrderCard extends RelativeLayout {
         @Override
         public void onRequested(int workOrderId) {
             if (_workOrder.getWorkOrderId() == workOrderId)
-                WorkOrderTracker.onActionButtonEvent(App.get(), _savedSearchTitle + " Saved Search", WorkOrderTracker.ActionButton.REQUEST, WorkOrderTracker.Action.REQUEST, (int) workOrderId);
+                WorkOrderTracker.onActionButtonEvent(App.get(), _savedSearchTitle + " Saved Search", WorkOrderTracker.ActionButton.REQUEST, WorkOrderTracker.Action.REQUEST, workOrderId);
         }
     };
 
@@ -637,7 +637,7 @@ public class WorkOrderCard extends RelativeLayout {
         @Override
         public void onAccepted(int workOrderId) {
             if (_workOrder.getWorkOrderId() == workOrderId)
-                WorkOrderTracker.onActionButtonEvent(App.get(), _savedSearchTitle + " Saved Search", WorkOrderTracker.ActionButton.ACCEPT_WORK, WorkOrderTracker.Action.ACCEPT_WORK, (int) workOrderId);
+                WorkOrderTracker.onActionButtonEvent(App.get(), _savedSearchTitle + " Saved Search", WorkOrderTracker.ActionButton.ACCEPT_WORK, WorkOrderTracker.Action.ACCEPT_WORK, workOrderId);
         }
     };
